+++ UIDbUpdateForm.py
@@ -18,7 +18,7 @@ class Ui_DbUpdateForm(object):
         DbUpdateForm.setMinimumSize(QtCore.QSize(450, 484))
         DbUpdateForm.setMaximumSize(QtCore.QSize(450, 484))
         self.groupInput_Files = QtWidgets.QGroupBox(DbUpdateForm)
-        self.groupInput_Files.setGeometry(QtCore.QRect(10, 190, 428, 169))
+        self.groupInput_Files.setGeometry(QtCore.QRect(10, 130, 428, 241))
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Expanding)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(5)
@@ -66,7 +66,7 @@ class Ui_DbUpdateForm(object):
         self.horizontalLayout_5.addWidget(self.comboFileType)
         self.verticalLayout_3.addLayout(self.horizontalLayout_5)
         self.groupInput_Files_2 = QtWidgets.QGroupBox(DbUpdateForm)
-        self.groupInput_Files_2.setGeometry(QtCore.QRect(10, 10, 428, 169))
+        self.groupInput_Files_2.setGeometry(QtCore.QRect(10, 10, 428, 111))
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Expanding)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(5)
